+++ src/main/java/woowacourse/shoppingcart/dto/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product;
 
 import woowacourse.shoppingcart.domain.Product;
