+++ src/main/java/woowacourse/shoppingcart/dto/product/ProductsRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product;
 
 public class ProductsRequest {
